third_party/blink/web_tests/svg/dom/document-createEvent-mandatory-arg-expected.txt

Tests that Document.createEvent() throws for SVG Documents when given no argument.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS svgDoc.__proto__ is XMLDocument.prototype
PASS svgDoc.__proto__.__proto__ is Document.prototype
PASS svgDoc.createEvent() threw exception TypeError: Failed to execute 'createEvent' on 'Document': 1 argument required, but only 0 present..
PASS svgDoc.createEvent('MouseEvents') did not throw exception.
PASS successfullyParsed is true

TEST COMPLETE
